# pkg/server/delete_analysistemplate_v1alpha1.go
package server

import (
	"context"
	"errors"
	"fmt"
	"net/http"

	"connectrpc.com/connect"
	"github.com/gin-gonic/gin"
	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"

	svcv1alpha1 "github.com/akuity/kargo/api/service/v1alpha1"
	"github.com/akuity/kargo/api/stubs/rollouts/v1alpha1"
)

func (s *server) DeleteAnalysisTemplate(
	ctx context.Context,
	req *connect.Request[svcv1alpha1.DeleteAnalysisTemplateRequest],
) (*connect.Response[svcv1alpha1.DeleteAnalysisTemplateResponse], error) {
	if !s.cfg.RolloutsIntegrationEnabled {
		// nolint:staticcheck
		return nil, connect.NewError(
			connect.CodeUnimplemented,
			errors.New("Argo Rollouts integration is not enabled"),
		)
	}

	project := req.Msg.GetProject()
	if err := validateFieldNotEmpty("project", project); err != nil {
		return nil, err
	}

	name := req.Msg.GetName()
	if err := validateFieldNotEmpty("name", name); err != nil {
		return nil, err
	}

	if err := s.validateProjectExists(ctx, project); err != nil {
		return nil, err
	}

	if err := s.client.Delete(ctx, &v1alpha1.AnalysisTemplate{
		ObjectMeta: metav1.ObjectMeta{
			Namespace: project,
			Name:      name,
		},
	}); err != nil {
		return nil, fmt.Errorf("delete AnalysisTemplate: %w", err)
	}

	return connect.NewResponse(&svcv1alpha1.DeleteAnalysisTemplateResponse{}), nil
}

// @id DeleteAnalysisTemplate
// @Summary Delete an AnalysisTemplate
// @Description Delete an AnalysisTemplate resource from a project's namespace.
// @Tags Verifications, Project-Level
// @Security BearerAuth
// @Param project path string true "Project name"
// @Param analysis-template path string true "AnalysisTemplate name"
// @Success 204 "Deleted successfully"
// @Router /v1beta1/projects/{project}/analysis-templates/{analysis-template} [delete]
func (s *server) deleteAnalysisTemplate(c *gin.Context) {
	if !s.cfg.RolloutsIntegrationEnabled {
		_ = c.Error(errArgoRolloutsIntegrationDisabled)
		return
	}

	ctx := c.Request.Context()

	project := c.Param("project")
	name := c.Param("analysis-template")

	if err := s.client.Delete(ctx, &v1alpha1.AnalysisTemplate{
		ObjectMeta: metav1.ObjectMeta{
			Namespace: project,
			Name:      name,
		},
	}); err != nil {
		_ = c.Error(err)
		return
	}

	c.Status(http.StatusNoContent)
}