# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Kargo",
        "title": "Kargo API",
        "contact": {},
        "version": "v1alpha1"
    },
    "basePath": "/",
    "paths": {
...
        "/v1beta1/projects/{project}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a Project resource.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core",
                    "Cluster-Scoped Resource"
                ],
                "summary": "Retrieve a Project resource",
                "operationId": "GetProject",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project custom resource (github.com/akuity/kargo/api/v1alpha1.Project)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a Project resource and its associated namespace.",
                "tags": [
                    "Core",
                    "Cluster-Scoped Resource"
                ],
                "summary": "Delete a Project",
                "operationId": "DeleteProject",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                }
            }
        },
...
        "/v1beta1/projects/{project}/analysis-templates/{analysis-template}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve an AnalysisTemplate resource from a project's\nnamespace.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verifications",
                    "Project-Level"
                ],
                "summary": "Retrieve an AnalysisTemplate",
                "operationId": "GetAnalysisTemplate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "AnalysisTemplate name",
                        "name": "analysis-template",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AnalysisTemplate custom resource (github.com/argoproj/argo-rollouts/pkg/apis/rollouts/v1alpha1.AnalysisTemplate)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an AnalysisTemplate resource from a project's namespace.",
                "tags": [
                    "Verifications",
                    "Project-Level"
                ],
                "summary": "Delete an AnalysisTemplate",
                "operationId": "DeleteAnalysisTemplate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "AnalysisTemplate name",
                        "name": "analysis-template",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                }
            }
        },
...
        "/v1beta1/resources": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update (or optionally, upsert) one or more Kargo resources from",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Update resources",
                "operationId": "UpdateResource",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "If true, create a resource if it does not exist",
                        "name": "upsert",
                        "in": "query"
                    },
                    {
                        "description": "YAML or JSON manifest(s)",
                        "name": "manifest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update results",
                        "schema": {
                            "$ref": "#/definitions/CreateOrUpdateResourceResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create one or more Kargo resources from YAML or JSON manifests.",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Create resources",
                "operationId": "CreateResource",
                "parameters": [
                    {
                        "description": "YAML or JSON manifest(s)",
                        "name": "manifest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "schema": {
                            "$ref": "#/definitions/CreateResourceResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete one or more Kargo resources using namespaces and names\nobtained from YAML or JSON manifests.",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Delete resources",
                "operationId": "DeleteResource",
                "parameters": [
                    {
                        "description": "YAML or JSON manifest(s)",
                        "name": "manifest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DeleteResourceResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateOrUpdateResourceResponse": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CreateOrUpdateResourceResult"
                    }
                }
            }
        },
        "CreateOrUpdateResourceResult": {
            "type": "object",
            "properties": {
                "createdResourceManifest": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "error": {
                    "type": "string"
                },
                "updatedResourceManifest": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
...
        "CreateResourceResponse": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CreateResourceResult"
                    }
                }
            }
        },
        "CreateResourceResult": {
            "type": "object",
            "properties": {
                "createdResourceManifest": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "DeleteResourceResponse": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DeleteResourceResult"
                    }
                }
            }
        },
        "DeleteResourceResult": {
            "type": "object",
            "properties": {
                "deletedResourceManifest": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication. Obtain token via OIDC/PKCE flow with your identity provider.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}